test(core): split build options exposure tests
